requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
imagekitio==4.1.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
requests-toolbelt==0.10.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==1.26.20
whitenoise==6.10.0
